Refactor character show presenter methods
