+++ app/models/character_show_text.rb
@@ -1,0 +1,15 @@
+module CharacterShowText
+
+  def title
+    "CHARACTER PROFILE"
+  end
+
+  def relation_1
+    "CHARACTER SERIES"
+  end
+
+  def relation_2
+    "CHARACTER EVENTS"
+  end
+
+end
